+++ compiler/src/lib.rs
@@ -1,8 +1,4 @@
-pub mod check;
 pub mod model;
 pub mod native;
 pub mod native_input;
 pub mod native_tool;
-pub mod package;
-pub mod parser;
-mod proof;